1.0.17 Fixed reloading issues Fixed /sellprice command

+++ src/main/java/me/drawethree/ultraprisoncore/autosell/api/UltraPrisonAutoSellAPI.java
@@ -8,7 +8,7 @@ public interface UltraPrisonAutoSellAPI {
 
     double getCurrentEarnings(Player player);
 
-    long getPriceForBrokenBlock(IWrappedRegion region, Block block);
+    double getPriceForBrokenBlock(IWrappedRegion region, Block block);
 
     boolean hasAutoSellEnabled(Player p);
 }
